+++ packages/fliphub-chain/fusebox-chain/Chainable.js
@@ -1,6 +1,7 @@
 module.exports = class {
   constructor(parent) {
-    this.parent = parent
+    this.parent = parent || this
+    if (this.init) this.init(this.parent)
   }
 
   end() {