+++ packages/fliphub2/bundle.js
@@ -1,0 +1,1 @@
+"use strict";const FlipBox=require("./src"),flip=FlipBox.base({root:__dirname,output:"dist/$name",sourceMaps:!0});console.log(flip);