+++ packages/fliphub2/src/index.js
@@ -1,0 +1,13 @@
+// const FlipBox = require('./core/FlipBox')
+// const cli = require('./hubs/CliHub')
+// const paths = require('./paths')
+// const es5exports = require('es5exports')
+//
+// exports.paths = paths
+// exports.cli = cli
+// exports.helpers = FlipBox.helpers
+// exports.flags = FlipBox.flags
+// exports['default'] = FlipBox
+//
+// module.exports = es5exports(exports['default'], exports)
+// // module.exports = FlipBox