+++ packages/fliphub2/src/presets/PresetReplace.js
@@ -1,0 +1,8 @@
+// @TODO:
+// use this, however important to note that
+// `rollup` + `fusebox` would use this,
+// yet webpack's `replace` is not the same
+//
+// maybe make my own better webpack replace
+// that runs before everything and has an easy api
+module.exports = class PresetReplace {}