+++ packages/fliphub2/src/presets/next/PresetsBabel.js
@@ -1,0 +1,1 @@
+https://github.com/rollup/rollup/issues/844