+++ packages/playground/original/middleware/-simpo/tasks/index.js
@@ -1,6 +1,0 @@
-var pkg = require('./pkg')
-var exportee = {
-  pkg,
-}
-
-module.exports = exportee