+++ packages/playground/original/middleware/-simpo/tasks/tests.js
@@ -1,4 +1,0 @@
-// mocha --require source-map-support/register /tmp/testBundle.js
-module.exports = function(app, helpers) {
-  return app
-}