+++ packages/playground/original/middleware/pkgBuilder.js
@@ -1,1 +1,0 @@
-// todo, for writing pieces out from scripts /packagejson to commands to logs